Fix to support both Custom Fields and AnyProps.
