+++ packages/formik/src/index.tsx
@@ -1,12 +1,15 @@
-export * from './hooks/useTypedField';
-export * from './hooks/useTypedFieldArray';
+
 export * from './hooks/useFormikState';
 export * from './hooks/hooks';
 export * from './Formik';
 export * from './Field';
+export * from './Field.helpers';
+export * from './Field.types';
+export * from './FieldArray';
+export * from './FieldArray.helpers';
+export * from './FieldArray.types';
 export * from './Form';
 export * from './withFormik';
-export * from './FieldArray';
 export * from './utils';
 export * from './types';
 export * from './connect';
