+++ packages/formik/test/type-checks.tsx
@@ -10,10 +10,10 @@ import {
     PathMatchingValue,
     RecursivelyTupleKeys,
     FieldComponentClass,
-    createTypedField,
     FieldConfig,
     CustomField,
-    useTypedField
+    useTypedField,
+    useCustomField
 } from "../src";
 
 type BasePerson = {
@@ -32,16 +32,9 @@ type Person = BasePerson & {
   friends: BasePerson[];
 }
 
-const TypedField = createTypedField<Person>();
-
 const proplessFC = () => null;
 const propsAnyFC = (props: any) => props ? null : null;
 class PropsAnyClass extends React.Component<any> {
-  render() {
-    return this.props.what ? null : null;
-  }
-}
-class ClassComponentWithOnlyExtra extends React.Component<{ what: true }> {
   render() {
     return this.props.what ? null : null;
   }
@@ -56,6 +49,9 @@ const renderAnyFn: FieldRenderFunction<any, any> =
 const asNumberFC: TypedAsField<number> =
   (props) => props.value ? null : null;
 
+const asStringFC: TypedAsField<string> =
+  (props) => props.value ? null : null;
+
 class AsNumberClass<Values> extends React.Component<FieldAsProps<number, Values>> {
   render() {
     return this.props.value ? null : null;
@@ -63,28 +59,13 @@ class AsNumberClass<Values> extends React.Component<FieldAsProps<number, Values>
 }
 const componentNumberFC: TypedComponentField<number> =
   (props) => props.field ? null : null;
-const componentNumberFCWithExtra: TypedComponentField<number> =
-  (props) => props.field ? null : null;
 const partialPropsFC = (props: {name: string}) => props.name ? null : null;
-const overlappingPropsFCWithExtra = (props: {name: string, what: true}) => props.name && props.what ? null : null;
 class ComponentNumberClass extends FieldComponentClass<number> {
   render() {
     return this.props.value ? null : null;
   }
 }
-class ComponentNumberClassWithExtra<Values> extends React.Component<
-  FieldComponentProps<number, Values>
-> {
-  render() {
-    return this.props.value ? null : null;
-  }
-}
 class PartialPropsClass extends React.Component<{name: string}> {
-  render() {
-    return this.props.name ? null : null;
-  }
-}
-class OverlappingPropsClassWithExtra extends React.Component<{name: string, what: true}> {
   render() {
     return this.props.name ? null : null;
   }
@@ -104,11 +85,23 @@ const CustomNumberFC: CustomField<number> = <Values,>(
       name={props.name}
       as={asNumberFC}
     />
+    <InnerTypedField
+      name={props.name}
+      as={asAnyFC}
+    />
+    <InnerTypedField
+      name={props.name}
+      /** @ts-expect-error value must match `number` */
+      as={asStringFC}
+    />
   </>
 }
 
-const fieldTests = (props: FieldConfig<"age", Person>) =>
-  <>
+const fieldTests = (props: FieldConfig<"age", Person>) => {
+  const TypedField = useTypedField<Person>();
+  const TypedNumberFC = useCustomField<Person>()(CustomNumberFC);
+
+  return <>
     {/* Default */}
     <Field name="age" />
     <TypedField name="age" />
@@ -211,8 +204,39 @@ const fieldTests = (props: FieldConfig<"age", Person>) =>
 
     {/* Pass-Through Props */}
     <TypedField {...props} />
-  </>
 
+    {/* Custom Fields */}
+    {/* Untyped Custom Fields can have anything */}
+    <CustomNumberFC name="age" format={value => {}} />
+    <TypedNumberFC name="age" format={value => {}} />
+    <CustomNumberFC name="motto" />
+    {/* @ts-expect-error field value should match */}
+    <TypedNumberFC name="motto" />
+  </>
+}
+
+// Not implemented
+//
+// class ClassComponentWithOnlyExtra extends React.Component<{ what: true }> {
+//   render() {
+//     return this.props.what ? null : null;
+//   }
+// }
+// const componentNumberFCWithExtra: TypedComponentField<number> =
+// (props) => props.field ? null : null;
+// const overlappingPropsFCWithExtra = (props: {name: string, what: true}) => props.name && props.what ? null : null;
+// class ComponentNumberClassWithExtra<Values> extends React.Component<
+//   FieldComponentProps<number, Values>
+// > {
+//   render() {
+//     return this.props.value ? null : null;
+//   }
+// }
+// class OverlappingPropsClassWithExtra extends React.Component<{name: string, what: true}> {
+//   render() {
+//     return this.props.name ? null : null;
+//   }
+// }
 // const featuresNotImplemented = () =>
 //   <>
 //     {/*
@@ -242,10 +266,6 @@ const fieldTests = (props: FieldConfig<"age", Person>) =>
 //     {/* @ts-expect-error propless components don't have extraProps */}
 //     <TypedField name="age" component={proplessFC} what={true} />
 
-//     {/* @ts-expect-error class component should match type */}
-//     <TypedField name="motto" as={AsNumberClass} />
-//     {/* @ts-expect-error class component should match type */}
-//     <TypedField name="motto" component={ComponentNumberClass} />
 
 //     {/*
 //       * Gave up on ExtraProps for now
@@ -351,116 +371,58 @@ const recursivelyTupledKeys: RecursivelyTupleKeys<Person>[] = [
   ["friends", 0, "favoriteNumbers", 0]
 ];
 
-const str: ValueMatchingPath<Person, "motto"> = person.motto;
-const strs: ValueMatchingPath<Person, "nicknames"> = person.nicknames;
-const str1: ValueMatchingPath<Person, "nicknames.1"> = person.nicknames[1];
-const num: ValueMatchingPath<Person, "age"> = person.age;
-const nums: ValueMatchingPath<Person, "favoriteNumbers"> = person.favoriteNumbers;
-const num1: ValueMatchingPath<Person, "favoriteNumbers.1"> = person.favoriteNumbers[1];
-const obj: ValueMatchingPath<Person, "partner"> = person.partner;
-const objstr: ValueMatchingPath<Person, "partner.motto"> = person.partner.motto;
-const objstrs: ValueMatchingPath<Person, "partner.nicknames"> = person.partner.nicknames;
-const objstr1: ValueMatchingPath<Person, "partner.nicknames.1"> = person.partner.nicknames[1];
-const objnum: ValueMatchingPath<Person, "partner.age"> = person.partner.age;
-const objnums: ValueMatchingPath<Person, "partner.favoriteNumbers"> = person.partner.favoriteNumbers;
-const objnum1: ValueMatchingPath<Person, "partner.favoriteNumbers.1"> = person.partner.favoriteNumbers[1];
-const obj1: ValueMatchingPath<Person, "friends.1"> = person.friends[1];
-const obj1str: ValueMatchingPath<Person, "friends.1.motto"> = person.friends[1].motto;
-const obj1num: ValueMatchingPath<Person, "friends.1.age"> = person.friends[1].age;
-const obj1nums: ValueMatchingPath<Person, "friends.1.favoriteNumbers"> = person.friends[1].favoriteNumbers;
-const obj1num1: ValueMatchingPath<Person, "friends.1.favoriteNumbers.1"> = person.friends[1].favoriteNumbers[1];
-
-const strMatches: PathMatchingValue<Person, string> = "motto";
-const strsMatches: PathMatchingValue<Person, string[]> = "nicknames";
-const str1Matches: PathMatchingValue<Person, string> = "nicknames.1";
-const numMatches: PathMatchingValue<Person, number> = "age";
-const numsMatches: PathMatchingValue<Person, number[]> = "favoriteNumbers";
-const num1Matches: PathMatchingValue<Person, number> = "favoriteNumbers.1";
-const objMatches: PathMatchingValue<Person, BasePerson> = "partner";
-const objstrMatches: PathMatchingValue<Person, string> = "partner.motto";
-const objstrsMatches: PathMatchingValue<Person, string[]> = "partner.nicknames";
-const objstr1Matches: PathMatchingValue<Person, string> = "partner.nicknames.1";
-const objnumMatches: PathMatchingValue<Person, number> = "partner.age";
-const objnumsMatches: PathMatchingValue<Person, number[]> = "partner.favoriteNumbers";
-const objnum1Matches: PathMatchingValue<Person, number> = "partner.favoriteNumbers.1";
-const obj1Matches: PathMatchingValue<Person, BasePerson> = "friends.1";
-const obj1strMatches: PathMatchingValue<Person, string> = "friends.1.motto";
-const obj1numMatches: PathMatchingValue<Person, number> = "friends.1.age";
-const obj1numsMatches: PathMatchingValue<Person, number[]> = "friends.1.favoriteNumbers";
-const obj1num1Matches: PathMatchingValue<Person, number> = "friends.1.favoriteNumbers.1";
+const strMatches: PathMatchingValue<string, Person> = "motto";
+const strsMatches: PathMatchingValue<string[], Person> = "nicknames";
+const str1Matches: PathMatchingValue<string, Person> = "nicknames.1";
+const numMatches: PathMatchingValue<number, Person> = "age";
+const numsMatches: PathMatchingValue<number[], Person> = "favoriteNumbers";
+const num1Matches: PathMatchingValue<number, Person> = "favoriteNumbers.1";
+const objMatches: PathMatchingValue<BasePerson, Person> = "partner";
+const objstrMatches: PathMatchingValue<string, Person> = "partner.motto";
+const objstrsMatches: PathMatchingValue<string[], Person> = "partner.nicknames";
+const objstr1Matches: PathMatchingValue<string, Person> = "partner.nicknames.1";
+const objnumMatches: PathMatchingValue<number, Person> = "partner.age";
+const objnumsMatches: PathMatchingValue<number[], Person> = "partner.favoriteNumbers";
+const objnum1Matches: PathMatchingValue<number, Person> = "partner.favoriteNumbers.1";
+const obj1Matches: PathMatchingValue<BasePerson, Person> = "friends.1";
+const obj1strMatches: PathMatchingValue<string, Person> = "friends.1.motto";
+const obj1numMatches: PathMatchingValue<number, Person> = "friends.1.age";
+const obj1numsMatches: PathMatchingValue<number[], Person> = "friends.1.favoriteNumbers";
+const obj1num1Matches: PathMatchingValue<number, Person> = "friends.1.favoriteNumbers.1";
 
 // @ts-expect-error
-const strValueFails: ValueMatchingPath<Person, "motto"> = person.age;
+const strFails: PathMatchingValue<number, Person> = "motto";
 // @ts-expect-error
-const strsValueFails: ValueMatchingPath<Person, "nicknames"> = person.favoriteNumbers;
+const strsFails: PathMatchingValue<number, Person[]> = "nicknames";
 // @ts-expect-error
-const str1ValueFails: ValueMatchingPath<Person, "nicknames.1"> = person.favoriteNumbers[0]
+const str1Fails: PathMatchingValue<number, Person> = "nicknames.1";
 // @ts-expect-error
-const numValueFails: ValueMatchingPath<Person, "age"> = person.motto;
+const numFails: PathMatchingValue<string, Person> = "age";
 // @ts-expect-error
-const numsValueFails: ValueMatchingPath<Person, "favoriteNumbers"> = person.nicknames;
+const numsFails: PathMatchingValue<string[], Person> = "favoriteNumbers";
 // @ts-expect-error
-const num1ValueFails: ValueMatchingPath<Person, "favoriteNumbers.1"> = person.nicknames[0];
-// @ts-expect-error
-const objValueFails: ValueMatchingPath<Person, "partner"> = person.friends;
-// @ts-expect-error
-const objstrValueFails: ValueMatchingPath<Person, "partner.motto"> = person.partner.age;
-// @ts-expect-error
-const objstrsValueFails: ValueMatchingPath<Person, "partner.nicknames"> = person.partner.favoriteNumbers;
-// @ts-expect-error
-const objstr1ValueFails: ValueMatchingPath<Person, "partner.nicknames.1"> = person.partner.favoriteNumbers[0]
-// @ts-expect-error
-const objnumValueFails: ValueMatchingPath<Person, "partner.age"> = person.partner.motto;
-// @ts-expect-error
-const objnumsValueFails: ValueMatchingPath<Person, "partner.favoriteNumbers"> = person.partner.nicknames;
-// @ts-expect-error
-const objnum1ValueFails: ValueMatchingPath<Person, "partner.favoriteNumbers.1"> = person.partner.nicknames[1];
-// @ts-expect-error
-const obj1ValueFails: ValueMatchingPath<Person, "friends.1"> = person.friends;
-// @ts-expect-error
-const obj1strValueFails: ValueMatchingPath<Person, "friends.1.motto"> = person.friends[1].age;
-// @ts-expect-error
-const obj1numValueFails: ValueMatchingPath<Person, "friends.1.age"> = person.friends[1].motto;
-// @ts-expect-error
-const obj1numsValueFails: ValueMatchingPath<Person, "friends.1.favoriteNumbers"> = person.friends[1].nicknames;
-// @ts-expect-error
-const obj1num1ValueFails: ValueMatchingPath<Person, "friends.1.favoriteNumbers.1"> = person.friends[1].nicknames[1];
-
-type NumberArrayMatchesString = ValueMatchingPath<Person, "nicknames.1" | "partner.nicknames.1"> extends number ? true : false;
-
-// @ts-expect-error
-const strFails: PathMatchingValue<Person, number> = "motto";
-// @ts-expect-error
-const strsFails: PathMatchingValue<Person, number[]> = "nicknames";
-// @ts-expect-error
-const str1Fails: PathMatchingValue<Person, number> = "nicknames.1";
-// @ts-expect-error
-const numFails: PathMatchingValue<Person, string> = "age";
-// @ts-expect-error
-const numsFails: PathMatchingValue<Person, string[]> = "favoriteNumbers";
-// @ts-expect-error
-const num1Fails: PathMatchingValue<Person, string> = "favoriteNumbers.1";
+const num1Fails: PathMatchingValue<string, Person> = "favoriteNumbers.1";
 // @ts-expect-error
 const objFails: PathMatchingValue<Person, Person> = "partner";
 // @ts-expect-error
-const objstrFails: PathMatchingValue<Person, number> = "partner.motto";
+const objstrFails: PathMatchingValue<number, Person> = "partner.motto";
 // @ts-expect-error
-const objstrsFails: PathMatchingValue<Person, number[]> = "partner.nicknames";
+const objstrsFails: PathMatchingValue<number, Person[]> = "partner.nicknames";
 // @ts-expect-error
-const objstr1Fails: PathMatchingValue<Person, number> = "partner.nicknames.1";
+const objstr1Fails: PathMatchingValue<number, Person> = "partner.nicknames.1";
 // @ts-expect-error
-const objnumFails: PathMatchingValue<Person, string> = "partner.age";
+const objnumFails: PathMatchingValue<string, Person> = "partner.age";
 // @ts-expect-error
-const objnumsFails: PathMatchingValue<Person, string[]> = "partner.favoriteNumbers";
+const objnumsFails: PathMatchingValue<string[], Person> = "partner.favoriteNumbers";
 // @ts-expect-error
-const objnum1Fails: PathMatchingValue<Person, string> = "partner.favoriteNumbers.1";
+const objnum1Fails: PathMatchingValue<string, Person> = "partner.favoriteNumbers.1";
 // @ts-expect-error
 const obj1Fails: PathMatchingValue<Person, Person> = "friends.1";
 // @ts-expect-error
-const obj1strFails: PathMatchingValue<Person, number> = "friends.1.motto";
+const obj1strFails: PathMatchingValue<number, Person> = "friends.1.motto";
 // @ts-expect-error
-const obj1numFails: PathMatchingValue<Person, string> = "friends.1.age";
+const obj1numFails: PathMatchingValue<string, Person> = "friends.1.age";
 // @ts-expect-error
-const obj1numsFails: PathMatchingValue<Person, string[]> = "friends.1.favoriteNumbers";
+const obj1numsFails: PathMatchingValue<string[], Person> = "friends.1.favoriteNumbers";
 // @ts-expect-error
-const obj1num1Fails: PathMatchingValue<Person, string> = "friends.1.favoriteNumbers.1";
+const obj1num1Fails: PathMatchingValue<string, Person> = "friends.1.favoriteNumbers.1";
